requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
openai>=1.0.0
exa-py>=1.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
